+++ client/js/components/github/github-search.test.js
