+++ client/js/components/shared/error-block.test.js
@@ -20,9 +20,10 @@ describe('components', () => {
         it('should render error when given a an error', () => {
             const { enzymeWrapper } = setup('error');
             expect(enzymeWrapper.find('div.error-block h1').text()).toBe('😢');
-            expect(enzymeWrapper.find('div.error-block h2').text()).toBe('ajax error 404');
+            expect(enzymeWrapper.find('div.error-block h2').text()).toBe(
+                'ajax error 404'
+            );
         });
     });
 });
-
 // 😊
