+++ client/js/ducks/github.test.js
@@ -1,12 +1,45 @@
 import deepFreeze from 'deepfreeze';
-import gitHubReducer,
-    {
+import gitHubReducer, {
+    GET_GITHUB_USER,
     GET_GITHUB_USER_FAILURE,
     GET_GITHUB_USER_SUCCESS,
+    getGitHubUser,
+    getGitHubUserFailure,
+    getGitHubUserSuccess,
     INITIAL_STATE as initialState
-    } from 'ducks/github';
+} from 'ducks/github';
 
 const INITIAL_STATE = deepFreeze(initialState);
+
+// Actions
+describe('getGitHubUser() actions', () => {
+    it('should create an action with the given userId as the payload.', () => {
+        const userId = 'pixel-fusion';
+        const expectedAction = {
+            type: GET_GITHUB_USER,
+            payload: userId
+        };
+        expect(getGitHubUser('pixel-fusion')).toEqual(expectedAction);
+    });
+});
+
+describe('getGitHubUserSuccess() actions', () => {
+    it('should create an action.', () => {
+        const expectedAction = {
+            type: GET_GITHUB_USER_SUCCESS
+        };
+        expect(getGitHubUserSuccess()).toEqual(expectedAction);
+    });
+});
+
+describe('getGitHubUserFailure() actions', () => {
+    it('should create an action.', () => {
+        const expectedAction = {
+            type: GET_GITHUB_USER_FAILURE
+        };
+        expect(getGitHubUserFailure()).toEqual(expectedAction);
+    });
+});
 
 describe('GitHub reducer default', () => {
     it('should return the existing state when given an action that isnt handled by this reducer', () => {
