+++ package.json
@@ -52,14 +52,14 @@
     "stylelint": "^7.7.0"
   },
   "dependencies": {
-    "babel-jest": "^19.0.0",
+    "babel-jest": "^20.0.3",
     "babel-loader": "^6.4.1",
     "babel-polyfill": "^6.20.0",
     "deepfreeze": "^2.0.0",
-    "enzyme": "^2.7.1",
+    "enzyme": "^2.8.2",
     "front-end-scripts": "https://github.com/pixelfusion/front-end-scripts.git#ci-build-fix",
     "immutable": "^3.8.1",
-    "jest": "^19.0.2",
+    "jest": "^20.0.4",
     "normalize.css": "^5.0.0",
     "prop-types": "^15.5.10",
     "react": "^15.4.2",
